refactor: make genesis block generation logic reusable
